Slicing: add CreateHammockCFG
So that prerequisities are met for the computation of slicing.
Otherwise infinite loops might be sliced away and that is incorrect.
This is due to how post-dominators are defined. The only path is
to an assert is through the loop => let's rip the loop off as it is
the only path and you have to go there once if you want to reach the
end (i.e. assert) node.

Signed-off-by: Jiri Slaby <jirislaby@gmail.com>